Update patch set 4
Patch Set 4:

(6 comments)

When the sticky titles stick i.e are taken out from the normal flow of their container, the container size changes. Even though they set to a fixed size, that was measured when the title is inside, it seems there is still some change, and that's why the nav seems to be inaccurate.

But the nav notes the position before scrolling, but during scrolling, the sticky title containers'height change, and the target container will be at a different position.

Patch-set: 4
Label: Code-Review=0